Add the golden-days-base pack to the list. Refactor GitHub workflow
The golden-days-base pack is interesting to add to this repository
because it contains audio files, and testing how they are optimized by
PackSquash can be useful.

While at it, I've also refactored the GitHub workflow definition to use
a matrix, which greatly reduces its verbosity and repetition, and
updated the submodule support test to point to the master branch of the
official repo, as that feature was already merged.